destroyアクションにて不正利用を制限

+++ app/controllers/items_controller.rb
@@ -33,8 +33,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if user_signed_in? && (current_user.id == item.user_id)
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to action: :index 
+    end
   end
 
   def show
